requirements.txt
# ── Core Framework ────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1

# ── Async & Concurrency ───────────────────────────────────────────────────────
aiohttp==3.9.5
httpx==0.27.0
anyio==4.3.0

# ── Web Scraping ──────────────────────────────────────────────────────────────
beautifulsoup4==4.12.3
lxml==5.2.1
requests==2.31.0
playwright==1.43.0          # JS-rendered CMS pages
fake-useragent==1.5.1

# ── NLP / Transformers (HuggingFace) ─────────────────────────────────────────
transformers==4.41.1
torch==2.3.0
tokenizers==0.19.1
datasets==2.19.1
accelerate==0.30.0
sentencepiece==0.2.0
protobuf==4.25.3

# ── NER / Medical Models ──────────────────────────────────────────────────────
spacy==3.7.4
scispacy==0.5.4             # Scientific NLP pipelines
# en_core_sci_lg model: pip install https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_lg-0.5.4.tar.gz

# ── Multi-Agent Orchestration ─────────────────────────────────────────────────
pyautogen==0.2.25
openai==1.30.1

# ── Database ──────────────────────────────────────────────────────────────────
motor==3.4.0
pymongo==4.7.2
redis==5.0.4

# ── Scheduling & Workers ──────────────────────────────────────────────────────
apscheduler==3.10.4
celery==5.4.0
kombu==5.3.3

# ── Dashboard ─────────────────────────────────────────────────────────────────
streamlit==1.35.0
plotly==5.22.0
pandas==2.2.2
altair==5.3.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
loguru==0.7.2
tenacity==8.3.0
cachetools==5.3.3
numpy==1.26.4
scikit-learn==1.4.2
python-dateutil==2.9.0
rich==13.7.1
typer==0.12.3
